// src/app/models/confession.ts
export class Confession {
  id: string;
  title: string;
  details: string;
  createdBy: string;
  createdByID: string;
  dateCreated: string;
  dateModified: string;
  timeCreated: string;
  image: string;
  isAnonymous: boolean;
  date:string;
  time:string;
  totalLikes: number;
  totalDisLikes: number;
  isLike: boolean;
  isReact: boolean;
  dialogueTypeNo: number;
  countryID:string;
}